+++ Dockerfile
@@ -1,0 +1,14 @@
+FROM python:3.10.16-slim
+
+WORKDIR /app
+
+COPY requirements.txt .
+
+RUN pip install --no-cache-dir -r requirements.txt
+
+COPY app/ app/
+COPY deep_learning_model/ deep_learning_model/
+
+EXPOSE 8000
+
+CMD ["uvicorn", "app.main:app", "--reload", "--host", "0.0.0.0", "--port", "8000"]
