+++ requirements_dev.txt
@@ -1,1 +1,13 @@
 -r requirements.txt
+black==25.1.0
+flake8==7.1.1
+isort==6.0.0
+nbconvert==7.16.6
+pre-commit==4.1.0
+pylint==3.3.4
+kagglehub[pandas-datasets]==0.3.7
+numpy==2.2.3
+pandas==2.2.3
+matplotlib==3.10.0
+torchinfo==1.8.0
+wandb==0.19.6
